+++ ru.taximaxim.codekeeper.core.tests/src/main/resources/ru/taximaxim/codekeeper/core/depcies/add_type_usr_t1_new.sql
@@ -1,0 +1,5 @@
+CREATE TYPE public.t1 AS RANGE (
+	subtype = date,
+	subtype_diff = public.t1_subtype_diff,
+	multirange_type_name = public.t1_multirange
+);